refactor: address code review (#134)
- Use Password for API secrets
- Extract setup wizard module
- Move dialoguer to workspace deps
- Add startup timeout (60 attempts)
- Propagate dialoguer errors with ?
- Use matches!() for AppPhase check
- Extract spinner frames to const

Closes #134
