+++ Cargo.toml
@@ -49,6 +49,7 @@ prost = "0.13"
 prost-types = "0.13"
 ratatui = "0.29"
 crossterm = "0.28"
+dialoguer = "0.11"
 
 [workspace.lints.rust]
 unsafe_code = "deny"
@@ -76,7 +77,7 @@ chrono.workspace = true
 clap.workspace = true
 config.workspace = true
 futures-util.workspace = true
-dialoguer = "0.11"
+dialoguer.workspace = true
 dirs = "6"
 rara-agent.workspace = true
 rara-domain.workspace = true
