+++ src/main.rs
@@ -426,7 +426,7 @@ async fn run() -> error::Result<()> {
         }
         Command::Setup { interactive, action } => {
             if interactive {
-                run_setup_interactive().await?;
+                rara_trading::setup_wizard::run(run_setup_init, run_setup_validate).await?;
             } else if let Some(action) = action {
                 run_setup(action).await?;
             } else {
@@ -1773,186 +1773,6 @@ fn run_strategy_installed() -> error::Result<()> {
 // ---------------------------------------------------------------------------
 // Setup command handlers
 // ---------------------------------------------------------------------------
-
-/// Interactive guided setup — walks the user through init, account creation, and validation.
-async fn run_setup_interactive() -> error::Result<()> {
-    use dialoguer::Confirm;
-
-    eprintln!("=== rara-trading interactive setup ===\n");
-
-    // --- Step 1: Init config files ---
-    let config_path = paths::config_file();
-    let accounts_path = paths::accounts_file();
-    let config_exists = config_path.exists();
-    let accounts_exists = accounts_path.exists();
-
-    if config_exists && accounts_exists {
-        eprintln!("Config files already exist:");
-        eprintln!("  {}", config_path.display());
-        eprintln!("  {}", accounts_path.display());
-        let overwrite = Confirm::new()
-            .with_prompt("Overwrite existing config files?")
-            .default(false)
-            .interact()
-            .unwrap_or(false);
-        if overwrite {
-            run_setup_init(true)?;
-        }
-    } else {
-        eprintln!("Generating config files...");
-        run_setup_init(false)?;
-    }
-
-    // --- Step 2: Add accounts ---
-    loop {
-        let add_account = Confirm::new()
-            .with_prompt("Add a trading account?")
-            .default(true)
-            .interact()
-            .unwrap_or(false);
-
-        if !add_account {
-            break;
-        }
-
-        interactive_add_account()?;
-    }
-
-    // --- Step 3: Validate ---
-    let validate = Confirm::new()
-        .with_prompt("Run validation checks?")
-        .default(true)
-        .interact()
-        .unwrap_or(true);
-
-    if validate {
-        eprintln!("\nRunning validation...");
-        run_setup_validate().await?;
-    }
-
-    eprintln!("\n=== Setup complete ===");
-    Ok(())
-}
-
-/// Prompt the user interactively for account details and save to accounts.toml.
-fn interactive_add_account() -> error::Result<()> {
-    use dialoguer::{Confirm, Input, Select};
-    use rara_trading_engine::account_config::{
-        AccountConfig, BrokerConfig, CcxtBrokerConfig, PaperBrokerConfig,
-    };
-
-    let id: String = Input::new()
-        .with_prompt("Account ID")
-        .interact_text()
-        .expect("account ID prompt");
-
-    let broker_options = &["paper", "ccxt"];
-    let broker_idx = Select::new()
-        .with_prompt("Broker type")
-        .items(broker_options)
-        .default(0)
-        .interact()
-        .expect("broker prompt");
-    let broker_type = broker_options[broker_idx];
-
-    let label: String = Input::new()
-        .with_prompt("Label (human-readable name)")
-        .default(id.clone())
-        .interact_text()
-        .expect("label prompt");
-
-    let contracts_str: String = Input::new()
-        .with_prompt("Contracts (comma-separated, e.g. BTC-USDT,ETH-USDT)")
-        .default(String::new())
-        .interact_text()
-        .expect("contracts prompt");
-
-    let contracts: Vec<String> = contracts_str
-        .split(',')
-        .map(|s| s.trim().to_string())
-        .filter(|s| !s.is_empty())
-        .collect();
-
-    let broker_config = match broker_type {
-        "paper" => {
-            let fill_price_str: String = Input::new()
-                .with_prompt("Fill price (leave empty for market price)")
-                .default(String::new())
-                .interact_text()
-                .expect("fill price prompt");
-            let fill_price = fill_price_str.parse::<f64>().ok();
-            BrokerConfig::Paper(PaperBrokerConfig { fill_price })
-        }
-        "ccxt" => {
-            let exchange_options = &["binance", "bybit", "okx"];
-            let exchange_idx = Select::new()
-                .with_prompt("Exchange")
-                .items(exchange_options)
-                .default(0)
-                .interact()
-                .expect("exchange prompt");
-            let exchange = exchange_options[exchange_idx].to_string();
-
-            let sandbox = Confirm::new()
-                .with_prompt("Use sandbox/testnet?")
-                .default(true)
-                .interact()
-                .unwrap_or(true);
-
-            let api_key: String = Input::new()
-                .with_prompt("API key")
-                .default(String::new())
-                .interact_text()
-                .expect("api key prompt");
-
-            let secret: String = Input::new()
-                .with_prompt("API secret")
-                .default(String::new())
-                .interact_text()
-                .expect("secret prompt");
-
-            let passphrase: String = Input::new()
-                .with_prompt("Passphrase (OKX only, leave empty otherwise)")
-                .default(String::new())
-                .interact_text()
-                .expect("passphrase prompt");
-
-            BrokerConfig::Ccxt(CcxtBrokerConfig {
-                exchange,
-                sandbox,
-                api_key,
-                secret,
-                passphrase: if passphrase.is_empty() {
-                    None
-                } else {
-                    Some(passphrase)
-                },
-            })
-        }
-        _ => unreachable!(),
-    };
-
-    let mut cfg = accounts_config::load_accounts();
-
-    if cfg.accounts.iter().any(|a| a.id == id) {
-        eprintln!("Account \"{id}\" already exists, skipping.");
-        return Ok(());
-    }
-
-    cfg.accounts.push(AccountConfig {
-        id: id.clone(),
-        label: Some(label),
-        broker_config,
-        enabled: true,
-        contracts,
-    });
-
-    accounts_config::save_accounts(&cfg).context(IoSnafu)?;
-    eprintln!("Account \"{id}\" added successfully.\n");
-
-    Ok(())
-}
-
 
 async fn run_setup(action: SetupAction) -> error::Result<()> {
     match action {
